Fixed size_t conversion warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@58899 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/serial/member.cpp
@@ -397,7 +397,7 @@ CMemberInfo* CMemberInfo::SetSetFlag(const bool* setFlag)
 
 CMemberInfo* CMemberInfo::SetSetFlag(const Uint4* setFlag)
 {
-    unsigned offset = GetIndex()-1;
+    size_t offset = GetIndex()-1;
     m_SetFlagOffset = TPointerOffsetType(setFlag) + (offset/16)*4;
     m_BitSetMask = 3<<(offset%16*2);
     UpdateFunctions();
